requirements.txt
selenium==3.141.0
webdriver_manager==3.3.0
Faker==6.6.3
requests==2.24.0
Pillow==8.2.0
